Nuxt-Flask/backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
gevent==24.11.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pytest==8.3.5
python-dotenv==1.1.0
